+++ btd/issuer_test.go
