+++ btd/spendlist.go
