+++ btd/spendlist_test.go
